+++ src/test/java/com/societegenerale/aut/main/myproject/domain/DomainClassAnnotatedWithJson.java
@@ -1,0 +1,10 @@
+package com.societegenerale.aut.main.myproject.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties
+public class DomainClassAnnotatedWithJson {
+
+    private String someAttribute;
+
+}