+++ src/test/java/com/societegenerale/aut/main/myproject/infrastructure/InfraClassUsingConfig.java
@@ -1,0 +1,12 @@
+package com.societegenerale.aut.main.myproject.infrastructure;
+
+import com.societegenerale.aut.main.myproject.config.ConfigClass;
+
+public class InfraClassUsingConfig {
+
+    public InfraClassUsingConfig() {
+
+        ConfigClass configClassThatShouldNotBeHere=new ConfigClass();
+
+    }
+}