+++ src/test/java/com/societegenerale/aut/main/myproject/infrastructure/InfraClassUsingSpring.java
@@ -1,0 +1,13 @@
+package com.societegenerale.aut.main.myproject.infrastructure;
+
+import org.springframework.core.SpringProperties;
+
+public class InfraClassUsingSpring {
+
+    public InfraClassUsingSpring() {
+
+        //some dummy code using Spring..
+        SpringProperties.getProperty("someProperty");
+
+    }
+}